new name format on text and chart view
