+++ dashboard-ui/src/components/TextBasedResults/TextBasedResultsPage.jsx
@@ -14,6 +14,7 @@ import XLSX from 'sheetjs-style';
 import Select from 'react-select';
 import { useSelector } from 'react-redux';
 import NoSelection from './NoSelection';
+import { shortenAnswer, p2Attributes } from './util';
 
 const get_eval_name_numbers = gql`
     query getEvalIdsForAllScenarioResults{
@@ -32,105 +33,19 @@ const GET_PARTICIPANT_LOG = gql`
 
 const KEYS_WITHOUT_TIME = ['kdmas', 'group_targets', 'lowAlignmentData', 'highAlignmentData'];
 
-function shortenAnswer(answer) {
-    // shortens an answer so the whole thing is visible and understandable without taking up too much screen real estate
-    switch (answer) {
-        case 'I mostly considered the potential to save the most lives.':
-            return "Save the Most Lives";
-        case 'I considered the intent to cause harm and the degree of responsibility for the situation.':
-            return "Intent to Cause Harm/Responsibility";
-        case 'I primarily considered the likely medical impact and tried to save the most lives.':
-            return "Medical Impact/Save the Most Lives";
-        case 'I considered the intent to cause harm, the degree of responsibility for the situation, and the potential to save the most lives.':
-            return 'Intent to Cause Harm/Responsibility/Save the Most Lives';
-        case 'The patient further from the IED':
-            return "Patient further from IED"
-        case 'The patient closer to the IED':
-            return "Patient closer to IED";
-        case 'Perform vitals assessment on Casualty W':
-            return "Vitals: Casualty W";
-        case 'Perform vitals assessment on Casualty X':
-            return "Vitals: Casualty X";
-        case 'Perform vitals assessment on Casualty V':
-            return "Vitals: Casualty V";
-        case 'Assess Casualty W for internal bleeding':
-            return "Assess Casualty W";
-        case 'Assess Casualty X for pelvic fracture':
-            return "Assess Casualty X";
-        case 'Provide burn dressing with pain control':
-            return "Burn dressing and pain control";
-        case 'Plan to treat both by alternating bag valve mask (BVM) ':
-            return "Treat both with BVM";
-        case 'Do not treat and search for help':
-            return 'Search for help';
-        case 'Apply Pain Med/Aid Bag to Casualty U':
-            return 'Treat Casualty U: Pain Meds/Aid Bag';
-        case 'Call to find out if additional support is possible':
-            return 'Call for support';
-        case 'Move forward with evacuating without confirming care capabilities':
-            return 'Evacuate';
-        case 'Hold casualties to determine if they can be transported':
-            return 'Hold for transport';
-        case 'Transport Casualty V to local hospital':
-            return 'Casualty V: transport to local hospital';
-        case 'Apply Decompression Needle to Casualty W':
-            return 'Casualty W: Needle Treatment';
-        case 'Check respiration on Casualty W':
-            return 'Casualty W: Respiration';
-        case 'Not assessing local capabilities and move forward with evacuating':
-            return 'Evacuate';
-        case 'Carefully divide the burn kit to cover the worst burns on each casualty':
-            return 'Divide Burn Kit';
-        case "Hold casualties until you get updated info on local hospital's capabilities to manage the casualties":
-            return 'Hold until info on local hospital capabilities';
-        case "Hold casualties until you get updated info on the capabilities of the embassy":
-            return 'Hold until info on embassy capabilities';
-        case "I considered the likely medical impact and also the intent to cause harm.":
-            return "Medical Impact/Intent to Cause Harm";
-        case "I considered the likely medical impact and also the degree of responsibility for the situation.":
-            return "Medical Impact/Degree of Responsibility";
-        case "Transport both casualties to the local hospital despite hospital availability":
-            return "Transport both to hospital";
-        case "Hold casualties to determine if they can be transported to safehouse":
-            return "Hold for transport to safehouse";
-        case "I considered the intent to cause harm, the degree of responsibility for the situation, and helpful things each patient had done.":
-            return "Intent to Harm/Responsibility/Helpfulness of Patient";
-        case "I considered the intent to cause harm, the degree of responsibility for the situation, the helpful things each patient had done, and the fact that the patients were from different groups":
-            return "Intent/Responsibility/Helpfulness/Different Groups";
-        case "Provide burn dressing with pain control (Burn kit)":
-            return "Burn Kit";
-        case "Perform an escharotomy (minor surgical kit)":
-            return "Escharotomy";
-        case "Get info on local hospital's capabilities to manage the casualties":
-            return "Info on Local Hospital";
-        case "Get info on the capabilities of the embassy to hold casualties until evacuation":
-            return "Info on Embassy";
-        case "I mostly considered the fact that the patients were from different groups.":
-            return "Patients are from different groups";
-        case 'Treat Patient A':
-            return "Treat Patient A";
-        case 'Treat Patient B':
-            return "Treat Patient B";
-        default:
-            if (answer.includes("BVM")) {
-                return "Treat both with BVM";
-            }
-            if (answer.includes("until you get updated info on the capabilities of the embassy")) {
-                return 'Hold until info on embassy capabilities';
-            }
-            if (answer.toLowerCase().includes('because')) {
-                let substring = answer.substring(answer.toLowerCase().indexOf('because'));
-                return substring.charAt(0).toUpperCase() + substring.slice(1);
-            }
-            return answer;
-    }
-}
+const cleanTitle = (title, scenario) => {
+    const foundAttr = p2Attributes.find(attr => scenario && scenario.includes(attr));
+    const attrOrScenario = foundAttr || '';
+    const match = title.match(/(\d+)\s*$/);
+    const number = match ? match[1] : title;
+    return `Probe_${attrOrScenario}_${number}`;
+};
 
-const cleanTitle = (title) => {
+const cleanTitleLegacy = (title) => {
     return title.replace(/probe\s*/, '').trim();
 };
 
-function SingleGraph({ data, pageName }) {
+function SingleGraph({ data, pageName, scenario, selectedEval }) {
     const [survey, setSurvey] = React.useState(null);
     const [vizPanel, setVizPanel] = React.useState(null);
     const [surveyResults, setSurveyResults] = React.useState([]);
@@ -200,28 +115,34 @@ function SingleGraph({ data, pageName }) {
 
     return (
         <div className='graph-section'>
-            <h3 className='question-header'>{cleanTitle(pageName)} (N={data['total']})</h3>
+            <h3 className='question-header'>{(selectedEval >= 8 ? cleanTitle(pageName, scenario) : cleanTitleLegacy(pageName))} (N={data['total']})</h3>
             <div id={"viz_" + pageName} className='full-width-graph' />
         </div>
     );
 }
 
-
 function getQuestionText(qkey, scenario, textBasedConfigs) {
+    console.log(qkey);
+    console.log(scenario);
+    console.log(textBasedConfigs);
+    return qkey;
+}
+
+function getQuestionTextLegacy(qkey, scenario, textBasedConfigs) {
     const pagesForScenario = textBasedConfigs[scenario]['pages'];
     for (const page of pagesForScenario) {
         for (const res of page['elements']) {
             if (res['name'] === qkey && res['choices']) {
                 // set title name
-                return cleanTitle(res['title'] + ' - ' + (scenario.includes("Adept") && qkey.includes("3") ? (qkey + 'a') : qkey.includes("Follow Up") ? (qkey.split("Follow Up")[0] + "3b") : qkey));
+                return cleanTitleLegacy(res['title'] + ' - ' + (scenario.includes("Adept") && qkey.includes("3") ? (qkey + 'a') : qkey.includes("Follow Up") ? (qkey.split("Follow Up")[0] + "3b") : qkey));
             }
         }
     }
-    return cleanTitle(qkey);
+    return cleanTitleLegacy(qkey);
 }
 
 
-function ParticipantView({ data, scenarioName, textBasedConfigs }) {
+function ParticipantView({ data, scenarioName, textBasedConfigs, selectedEval }) {
     const [organizedData, setOrganizedData] = React.useState(null);
     const [excelData, setExcelData] = React.useState(null);
     const [orderedHeaders, setHeaders] = React.useState([]);
@@ -318,7 +239,7 @@ function ParticipantView({ data, scenarioName, textBasedConfigs }) {
                 <thead>
                     <tr>
                         {orderedHeaders.map((key) => {
-                            return <th key={scenarioName + "_" + key}>{getQuestionText(key, scenarioName, textBasedConfigs)}</th>
+                            return <th key={scenarioName + "_" + key}>{(selectedEval >= 8 ? getQuestionText(key, scenarioName, textBasedConfigs) : getQuestionTextLegacy(key, scenarioName, textBasedConfigs))}</th>
                         })}
                     </tr>
                 </thead>
@@ -364,7 +285,7 @@ export default function TextBasedResultsPage() {
     const [responsesByScenario, setByScenario] = React.useState(null);
     const [questionAnswerSets, setResults] = React.useState(null);
     const [participantBased, setParticipantBased] = React.useState(null);
-    const [selectedEval, setSelectedEval] = React.useState(5);
+    const [selectedEval, setSelectedEval] = React.useState(8);
     const [evalOptions, setEvalOptions] = React.useState([]);
     const [scenarioOptions, setScenarioOptions] = React.useState([]);
     const textBasedConfigs = useSelector(state => state.configs.textBasedConfigs);
@@ -550,7 +471,7 @@ export default function TextBasedResultsPage() {
                     .map((qkey, ind) => {
                         const displayKey = questionAnswerSets[qkey].originalKey || qkey;
                         return (<div className='result-section' key={displayKey + '_' + ind}>
-                            <h3 className='question-header'>{cleanTitle(displayKey)} (N={questionAnswerSets[qkey]['total']})</h3>
+                            <h3 className='question-header'>{(selectedEval >= 8 ? cleanTitle(displayKey, scenarioChosen) : cleanTitleLegacy(displayKey))} (N={questionAnswerSets[qkey]['total']})</h3>
                             <p>{questionAnswerSets[qkey]['question']}</p>
                             <table className="itm-table text-result-table">
                                 <thead>
@@ -600,7 +521,7 @@ export default function TextBasedResultsPage() {
                         return a.localeCompare(b);
                     })
                     .map((qkey) => {
-                        return (<SingleGraph key={qkey} data={questionAnswerSets[qkey]} pageName={qkey} />);
+                        return (<SingleGraph key={qkey} data={questionAnswerSets[qkey]} pageName={qkey} scenario={scenarioChosen} selectedEval={selectedEval} />);
                     })
                 : loading ? <h2 className="no-data">Loading Data...</h2> : <h2 className="no-data">No Data Found</h2>}
         </div>);
@@ -662,7 +583,7 @@ export default function TextBasedResultsPage() {
                         <ToggleButton variant="secondary" id='choose-participant' value={"participants"}>Participants</ToggleButton>
                     </ToggleButtonGroup>
                 </div>
-                {dataFormat === 'text' ? <TextResultsSection /> : dataFormat === 'participants' ? <ParticipantView data={scenarioChosen && participantBased ? participantBased[scenarioChosen] : []} scenarioName={scenarioChosen} textBasedConfigs={filteredTextBasedConfigs} /> : <ChartedResultsSection />}
+                {dataFormat === 'text' ? <TextResultsSection /> : dataFormat === 'participants' ? <ParticipantView data={scenarioChosen && participantBased ? participantBased[scenarioChosen] : []} scenarioName={scenarioChosen} textBasedConfigs={filteredTextBasedConfigs} selectedEval={selectedEval} /> : <ChartedResultsSection />}
             </div>) : (
             <NoSelection />
         )}
